embed_service/requirements.txt
sentence-transformers==2.7.0
grpcio==1.64.0
grpcio-tools==1.64.0
torch==2.3.0
flask==3.0.3
